+++ R/dbi-driver.R
@@ -42,12 +42,21 @@ setMethod("show", "OdbcDriver",
 #'   timezone that is _not_ 'UTC'. If the database is in your local timezone,
 #'   set this argument to [Sys.timezone()]. See [OlsonNames()] for a complete
 #'   list of available time zones on your system.
+#'   Note, if the datatype itself carries timezone information, as is the case,
+#'   for example, with SQL Server::DATETIMEOFFSET, `package:odbc` will make
+#'   an effort to respect the timezone declared therein.  In those cases,
+#'   this parameter is not used - the timezone that is part of the datatype
+#'   takes precedence.
 #' @param timezone_out The time zone returned to R. If you want to display
 #'   datetime values in the local timezone, set to [Sys.timezone()].
 #' @param encoding The text encoding used on the Database. If the database is
 #'   not using UTF-8 you will need to set the encoding to get accurate
 #'   re-encoding. See [iconvlist()] for a complete list of available encodings
 #'   on your system. Note strings are always returned `UTF-8` encoded.
+#' @param name_encoding The text encoding for column names used on the
+#'   Database.  May be different than the `encoding` argument.  Defaults to
+#'   empty string which is equivalent to returning the column names without
+#'   performing any conversion.
 #' @param driver The ODBC driver name or a path to a driver. For currently
 #'   available options, see the `name` column of [odbcListDrivers()] output.
 #' @param server The server hostname. Some drivers use `Servername` as the name
@@ -64,7 +73,7 @@ setMethod("show", "OdbcDriver",
 #'   the driver does not return a valid value, it can be set manually with this
 #'   parameter.
 #' @param attributes A list of connection attributes that are passed
-#'   prior to the connection being established. See \link{ConnectionAttributes}.
+#'   prior to the connection being established. See [ConnectionAttributes].
 #' @param interruptible Logical.  If `TRUE` calls to `SQLExecute` and
 #'   `SQLExecuteDirect` can be interrupted when the user sends SIGINT ( ctrl-c ).
 #'   Otherwise, they block.  Defaults to `TRUE` in interactive sessions, and
@@ -138,8 +147,9 @@ setMethod("show", "OdbcDriver",
 #'   driver to create a connection object.
 #' 3) *Interface with connections*: The resulting connection object can be
 #'   passed to various functions to retrieve information on database
-#'   structure ([dbListTables()]), iteratively develop queries ([dbSendQuery()],
-#'   [dbColumnInfo()]), and query data objects ([dbFetch()]).
+#'   structure ([DBI::dbListTables()][]), iteratively develop queries
+#'   ([DBI::dbSendQuery()], [DBI::dbColumnInfo()]), and query data objects
+#'   ([DBI::dbFetch()]).
 #'
 #' @aliases dbConnect odbc
 #'
@@ -164,6 +174,7 @@ setMethod("dbConnect", "OdbcDriver",
       timezone = "UTC",
       timezone_out = "UTC",
       encoding = "",
+      name_encoding = "",
       bigint = c("integer64", "integer", "numeric", "character"),
       timeout = 10,
       driver = NULL,
@@ -179,6 +190,7 @@ setMethod("dbConnect", "OdbcDriver",
     check_string(timezone)
     check_string(timezone_out)
     check_string(encoding)
+    check_string(name_encoding)
     arg_match(bigint)
     check_number_decimal(timeout, allow_na = TRUE)
     check_string(driver, allow_null = TRUE)
@@ -199,6 +211,7 @@ setMethod("dbConnect", "OdbcDriver",
       timezone = timezone,
       timezone_out = timezone_out,
       encoding = encoding,
+      name_encoding = name_encoding,
       bigint = bigint,
       timeout = timeout,
       driver = driver,