+++ R/utils.R
@@ -162,7 +162,7 @@ random_name <- function(prefix = "") {
 rethrow_database_error <- function(msg, call = trace_back()$call[[1]]) {
   tryCatch(
     res <- parse_database_error(msg),
-    error = function(e) cli::cli_abort(msg, call = call)
+    error = function(e) cli::cli_abort("{msg}", call = call)
   )
 
   cli::cli_abort(
@@ -198,6 +198,15 @@ parse_database_error <- function(msg) {
     cnd_msg[-1],
     right = TRUE
   )
+  # Limit error message length to something sane (#914)
+  max_len <- 5000
+  cnd_body <- sapply(cnd_body, FUN = function(x) {
+    if (nchar(x) > max_len) {
+      x <- substr(x, 0, max_len)
+      x <- paste0(x, "...")
+    }
+    x
+  }, USE.NAMES = FALSE)
 
   cnd_body <- contextualize_database_error(cnd_body)
 
@@ -232,6 +241,8 @@ set_database_error_names <- function(cnd_body) {
     return(cnd_body)
   }
 
+  cnd_body <- escape_curly_brackets(cnd_body)
+
   if (is.null(names(cnd_body))) {
     return(
       set_names(cnd_body, nm = c("x", rep("*", length(cnd_body) - 1)))
@@ -245,6 +256,19 @@ set_database_error_names <- function(cnd_body) {
       ifelse(names(cnd_body[-1]) == "", "*", names(cnd_body[-1]))
     )
   )
+}
+
+# Escape curly brackets before formatting with cli (#859). Do so only to
+# unnamed elements so that formatting is preserved for context
+# added with `contextualize_database_error()`.
+escape_curly_brackets <- function(cnd_body) {
+  if (is.null(names(cnd_body))) {
+    return(gsub("\\{", "{{", gsub("\\}", "}}", cnd_body)))
+  }
+
+  unnamed <- names(cnd_body) == ""
+  cnd_body[unnamed] <- gsub("\\{", "{{", gsub("\\}", "}}", cnd_body[unnamed]))
+  cnd_body
 }
 
 # check helpers for common odbc arguments --------------------------------------
@@ -288,8 +312,24 @@ check_attributes <- function(attributes, call = caller_env()) {
 }
 
 # apple + spark drive config (#651) --------------------------------------------
-configure_spark <- function(call = caller_env()) {
+# Method will attempt to:
+# 1. Locate an installation of unixodbc / error out otherwise.
+# 2. Verify the driver_config argument.  Expect this to be a list with
+#    two fields:
+#    * path Vector of viable driver paths ( only first one is used )
+#    * url A location where the user can downlaod the driver from.
+#    See spark_simba_config, for example.  Its return value is used as
+#    the value for this argument.
+# 3. Inspect the config for some settings that can impact how our package
+#    performs.
+# 4. If action == "modify" then we attempt to modify the config in-situ.
+# 5. Otherwise we throw a warning asking the user to revise.
+configure_simba <- function(driver_config,
+                            action = "modify", call = caller_env()) {
   if (!is_macos()) {
+    return(invisible())
+  }
+  if (!is.null(getOption("odbc.no_config_override"))) {
     return(invisible())
   }
 
@@ -298,18 +338,18 @@ configure_spark <- function(call = caller_env()) {
     error_install_unixodbc(call)
   }
 
-  spark_config <- locate_config_spark()
-  if (length(spark_config) == 0) {
-    abort(
-      c(
-        "Unable to locate the needed spark ODBC driver.",
-        i = "Please install the needed driver from https://www.databricks.com/spark/odbc-drivers-download."
-      ),
+  simba_config <- driver_config$path
+  if (length(simba_config) == 0) {
+    func <- cli::cli_warn
+    if (action == "modify") {
+      fun <- cli::cli_abort
+    }
+    func(
+      c(i = "Please install the needed driver from {driver_config$url}"),
       call = call
     )
   }
-
-  configure_unixodbc_spark(unixodbc_install[1], spark_config[1], call)
+  configure_unixodbc_simba(unixodbc_install[1], simba_config[1], action, call)
 }
 
 locate_install_unixodbc <- function() {
@@ -327,8 +367,12 @@ locate_install_unixodbc <- function() {
     "/usr/local",
     "/lib",
     "/usr/local/lib",
+    "/usr/lib/x86_64-linux-gnu",
     "/opt/homebrew/lib",
-    "/opt/homebrew/opt/unixodbc/lib"
+    "/opt/homebrew/etc",
+    "/opt/homebrew/opt/unixodbc/lib",
+    "/opt/R/arm64/lib",
+    "/opt/R/x86_64/lib"
   )
 
   list.files(
@@ -342,7 +386,12 @@ system_safely <- function(x) {
   tryCatch(
     {
       unixodbc_prefix <- system(x, intern = TRUE, ignore.stderr = TRUE)
-      return(paste0(unixodbc_prefix, "/", libodbcinst_filename()))
+      candidates <- list.files(unixodbc_prefix,
+        pattern = libodbcinst_filename(), full.names = TRUE)
+      if (!length(candidates)) {
+        stop("Unable to locate unixodbc using odbc_config")
+      }
+      return(candidates[1])
     },
     error = function(e) {},
     warning = function(w) {}
@@ -350,11 +399,13 @@ system_safely <- function(x) {
   character()
 }
 
+# Returns a pattern to be used with
+# list.files( ..., pattern = ... ).
 libodbcinst_filename <- function() {
   if (is_macos()) {
-    "libodbcinst.dylib"
+    "libodbcinst.dylib|libodbcinst.a"
   } else {
-    "libodbcinst.so"
+    "libodbcinst.so|libodbcinst.a"
   }
 }
 
@@ -369,51 +420,51 @@ error_install_unixodbc <- function(call) {
   )
 }
 
-# p. 44 https://downloads.datastax.com/odbc/2.6.5.1005/Simba%20Spark%20ODBC%20Install%20and%20Configuration%20Guide.pdf
-locate_config_spark <- function() {
-  spark_env <- Sys.getenv("SIMBASPARKINI")
-  if (!identical(spark_env, "")) {
-    return(spark_env)
-  }
+configure_unixodbc_simba <- function(unixodbc_install, simba_config, action, call) {
 
-  common_dirs <- c(
-    "/Library/simba/spark/lib",
-    "/etc",
-    getwd(),
-    Sys.getenv("HOME")
-  )
-
-  list.files(
-    common_dirs,
-    pattern = "simba\\.sparkodbc\\.ini$",
-    full.names = TRUE
-  )
-}
-
-configure_unixodbc_spark <- function(unixodbc_install, spark_config, call) {
   # As shipped, the simba spark ini has an incomplete final line
   suppressWarnings(
-    spark_lines <- readLines(spark_config)
+    simba_lines <- readLines(simba_config)
   )
-
-  spark_lines_new <- replace_or_append(
-    lines = spark_lines,
-    pattern = "^ODBCInstLib=",
+  res <- replace_or_append(
+    lines = simba_lines,
+    key_pattern = "^ODBCInstLib=",
+    accepted_value = unixodbc_install,
     replacement = paste0("ODBCInstLib=", unixodbc_install)
   )
-
-  spark_lines_new <- replace_or_append(
-    lines = spark_lines_new,
-    pattern = "^DriverManagerEncoding=",
+  warnings <- character()
+  if (action != "modify" && res$modified) {
+     warnings <- c(warnings, c("*" = "Please consider revising the {.arg ODBCInstLib}
+       field in {.file {simba_config}} and setting its value to {.val {unixodbc_install}}"))
+  }
+  simba_lines_new <- res$new_lines
+  res <- replace_or_append(
+    lines = simba_lines_new,
+    key_pattern = "^DriverManagerEncoding=",
+    accepted_value = "UTF-16|utf-16",
     replacement = "DriverManagerEncoding=UTF-16"
   )
+  if (action != "modify" && res$modified) {
+     warnings <- c(warnings, c("*" = "Please consider revising the
+       {.arg DriverManagerEncoding} field in {.file {simba_config}} and setting its
+       value to {.val UTF-16}."))
+  }
+  if (length(warnings)) {
+    cli::cli_warn(c(
+      c(i = "Detected potentially unsafe driver settings:"),
+      warnings
+    ))
+  }
+  simba_lines_new <- res$new_lines
 
-  write_spark_lines(spark_lines, spark_lines_new, spark_config, call)
+  if (action == "modify") {
+    write_simba_lines(simba_lines, simba_lines_new, simba_config, call)
+  }
 
   invisible()
 }
 
-write_spark_lines <- function(spark_lines, spark_lines_new, spark_config, call) {
+write_simba_lines <- function(spark_lines, spark_lines_new, spark_config, call) {
   if (identical(spark_lines, spark_lines_new)) {
     return(invisible())
   }
@@ -436,21 +487,84 @@ write_spark_lines <- function(spark_lines, spark_lines_new, spark_config, call) 
   writeLines(spark_lines_new, spark_config)
 }
 
+# Interpret the argument as an `ODBC` driver
+# and attempt to infer the directory housing it.
+# It will return an empty character vector if unable to.
+driver_dir <- function(driver) {
+  # driver argument could be an outright path, or a name
+  # of a driver specified in odbcinst.ini  Try to discern
+  driver_spec <- subset(odbcListDrivers(), name == driver)
+  if (nrow(driver_spec)) {
+    driver_path <- subset(driver_spec, attribute == "Driver")$value
+  } else {
+    driver_path <- driver
+  }
+
+  ret <- dirname(driver_path)
+  if (ret == ".") {
+    ret <- character()
+  }
+  return(ret)
+}
+
 is_writeable <- function(path) {
   tryCatch(file.access(path, mode = 2)[[1]] == 0, error = function(e) FALSE)
 }
 
-# given a vector of lines in an ini file, look for a given key pattern.
-# the `replacement` is the whole intended line, giving the "key=value" pair.
-# if the key is found, replace that line with `replacement`.
-# if the key isn't found, append a new line with `replacement`.
-replace_or_append <- function(lines, pattern, replacement) {
-  matching_lines_loc <- grepl(pattern, lines)
+# Given a vector of lines in an ini file, look for a given key pattern.
+# If found:
+# - No action if the `accepted_value` argument is found on line.
+# - Replace otherwise.
+# If not found: append.
+# The `replacement` is the whole intended line, giving the desired
+# "key=value" pair.
+# @return a list with two elements:
+# - new_lines: Potentially modified lines
+# - modified: Whether method modified lines, where modified means
+# both changed or appended.
+replace_or_append <- function(lines, key_pattern, accepted_value, replacement) {
+  matching_lines_loc <- grepl(key_pattern, lines)
   matching_lines <- lines[matching_lines_loc]
+  found_ok <- length(matching_lines) != 0 &&
+    any(grepl(accepted_value, lines[matching_lines_loc]))
   if (length(matching_lines) == 0) {
     lines <- c(lines, replacement)
-  } else {
+  } else if (!found_ok) {
     lines[matching_lines_loc] <- replacement
   }
-  lines
+  return(list(new_lines = lines, modified = !found_ok))
+}
+
+running_on_connect <- function() {
+  Sys.getenv("RSTUDIO_PRODUCT") == "CONNECT"
+}
+
+# Like Sys.getenv(), but checks passed keys in order and allows for non-string
+# defaults.
+getenv2 <- function(..., .unset = NULL) {
+  x <- as.character(list(...))
+  env <- Sys.getenv(x, NA_character_)
+  if (all(is.na(env))) {
+    return(.unset)
+  }
+  val <- env[!is.na(env)][1]
+  unname(val)
+}
+
+find_default_driver <- function(paths, fallbacks, label, call = caller_env()) {
+  paths <- paths[file.exists(paths)]
+  if (length(paths) > 0) {
+    return(paths[1])
+  }
+  fallbacks <- intersect(fallbacks, odbcListDrivers()$name)
+  if (length(fallbacks) > 0) {
+    return(fallbacks[1])
+  }
+  cli::cli_abort(
+    c(
+      "Failed to automatically find the {label} ODBC driver.",
+      i = "Set {.arg driver} to known driver name or path."
+    ),
+    call = call
+  )
 }