+++ tests/testthat/test-driver-postgres.R
@@ -1,3 +1,5 @@
+skip_if_no_unixodbc()
+
 test_that("PostgreSQL", {
   DBItest::make_context(
     odbc(),
@@ -57,7 +59,6 @@ test_that("PostgreSQL", {
     "unquote_identifier_vectorized", # TODO
     "create_table_overwrite", # TODO
     "create_table_error", # TODO
-    "create_temporary_table", # TODO
     "append_table_.*", # TODO
     "append_roundtrip_.*", # TODO
     "append_table_.*", # TODO
@@ -165,4 +166,3 @@ test_that("odbcPreviewObject", {
   })
   expect_equal(nrow(res), 3)
 })
-